Update sqlact.cpp: add casts for MSVC 64-bit warnings

+++ source/sqlact.cpp
@@ -112,7 +112,7 @@ int SqlParse(sql_stmt *stmt, const char *query, int queryLen)
     return 0;
   }
   if (!queryLen) {
-    queryLen = strlen(query) + 1;
+    queryLen = static_cast<int>(strlen(query)) + 1;
   }
 
   stmt->query = query;
@@ -351,7 +351,7 @@ HB_FUNC(SR_HEXTOSTR)
     return;
   }
 
-  outbuff = sr_Hex2Str(hb_parc(1), hb_parclen(1), &nalloc);
+  outbuff = sr_Hex2Str(hb_parc(1), static_cast<int>(hb_parclen(1)), &nalloc);
   hb_retclen_buffer(outbuff, nalloc);
 }
 
@@ -647,14 +647,14 @@ HB_FUNC(SR_ESCAPENUM)
   for (iPos = 0; iPos < iSize; iPos++) {
     if (FromBuffer[iPos] == ',') {
       ToBuffer[iPos] = '.';
-      iDecPos = iPos;
+      iDecPos = static_cast<int>(iPos);
     } else {
       ToBuffer[iPos] = FromBuffer[iPos];
     }
 
     if (ToBuffer[iPos] == '.') {
       bInteger = false;
-      iDecPos = iPos;
+      iDecPos = static_cast<int>(iPos);
     }
 
     if (ToBuffer[iPos] == 'E' && (iPos + 2) <= iSize) { // 1928773.3663E+003
@@ -715,14 +715,14 @@ HB_FUNC(SR_ESCAPENUM)
 
     if (!iOverflow) {
       auto dValue = static_cast<double>(lValue);
-      hb_retnlen(dValue, len, dec);
+      hb_retnlen(dValue, static_cast<int>(len), static_cast<int>(dec));
     } else {
       double dValue = hb_strVal(ToBuffer, iSize);
-      hb_retnlen(dValue, len, dec);
+      hb_retnlen(dValue, static_cast<int>(len), static_cast<int>(dec));
     }
   } else {
     double dValue = hb_strVal(ToBuffer, iSize);
-    hb_retnlen(dValue, len, dec);
+    hb_retnlen(dValue, static_cast<int>(len), static_cast<int>(dec));
   }
   hb_xfree(ToBuffer);
 }
@@ -749,14 +749,14 @@ PHB_ITEM sr_escapeNumber(char *FromBuffer, HB_SIZE len, HB_SIZE dec, PHB_ITEM pR
   for (iPos = 0; iPos < iSize; iPos++) {
     if (FromBuffer[iPos] == ',') {
       ToBuffer[iPos] = '.';
-      iDecPos = iPos;
+      iDecPos = static_cast<int>(iPos);
     } else {
       ToBuffer[iPos] = FromBuffer[iPos];
     }
 
     if (ToBuffer[iPos] == '.') {
       bInteger = false;
-      iDecPos = iPos;
+      iDecPos = static_cast<int>(iPos);
     }
 
     if (ToBuffer[iPos] == 'E' && (iPos + 2) <= iSize) { // 1928773.3663E+003
@@ -817,14 +817,14 @@ PHB_ITEM sr_escapeNumber(char *FromBuffer, HB_SIZE len, HB_SIZE dec, PHB_ITEM pR
 
     if (!iOverflow) {
       auto dValue = static_cast<double>(lValue);
-      hb_itemPutNLen(pRet, dValue, len, dec);
+      hb_itemPutNLen(pRet, dValue, static_cast<int>(len), static_cast<int>(dec));
     } else {
       double dValue = hb_strVal(ToBuffer, iSize);
-      hb_itemPutNLen(pRet, dValue, len, dec);
+      hb_itemPutNLen(pRet, dValue, static_cast<int>(len), static_cast<int>(dec));
     }
   } else {
     double dValue = hb_strVal(ToBuffer, iSize);
-    hb_itemPutNLen(pRet, dValue, len, dec);
+    hb_itemPutNLen(pRet, dValue, static_cast<int>(len), static_cast<int>(dec));
   }
   hb_xfree(ToBuffer);
   return pRet;
